+++ src/Plainion.GraphViz.Modules.Documents/OpenDocumentsViewModel.cs
@@ -159,6 +159,11 @@ namespace Plainion.GraphViz.Modules.Documents
 
                 var writer = new DotWriter(path);
                 writer.PrettyPrint = true;
+                if (p.GetModule<IGraphLayoutModule>().Algorithm == LayoutAlgorithm.Flow)
+                {
+                    writer.Settings = DotPresets.Flow;
+                }
+
                 writer.Write(graph, new NullGraphPicking(), p);
             }