+++ src/Plainion.GraphViz.Modules.Documents/SaveDocumentsViewModel.cs
@@ -117,6 +117,12 @@ namespace Plainion.GraphViz.Modules.Documents
         private void SaveAsDot(string path)
         {
             var writer = new DotWriter(path);
+
+            if (Model.Presentation.GetModule<IGraphLayoutModule>().Algorithm == LayoutAlgorithm.Flow)
+            {
+                writer.Settings = DotPresets.Flow;
+            }
+
             writer.Write(Model.Presentation.GetModule<ITransformationModule>().Graph, Model.Presentation.Picking, Model.Presentation);
         }