+++ src/Plainion.GraphViz/Dot/LayoutAlgorithm.cs
@@ -4,7 +4,8 @@ namespace Plainion.GraphViz.Dot
     public enum LayoutAlgorithm
     {
         Auto,
-        Dot,
+        Hierarchy,
+        Flow,
         Sfdp
     }
 }